@@ -1,12 +1,11 @@
-﻿namespace Peaky.Client
-{
-    public enum TestOutcome
-    {
-        Passed,
-        Failed,
-        Inconclusive,
-        Error,
-        Timeout
+﻿namespace Peaky.Client;
 
-    }
+public enum TestOutcome
+{
+    Passed,
+    Failed,
+    Inconclusive,
+    Error,
+    Timeout
+
 }
