@@ -1,15 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Peaky.Client
-{
-    public class TestsResponse
-    {
-        public IEnumerable<Test> Tests { get; }
+namespace Peaky.Client;
 
-        public TestsResponse(IEnumerable<Test> tests)
-        {
-            Tests = tests;
-        }
+public class TestsResponse
+{
+    public IEnumerable<Test> Tests { get; }
+
+    public TestsResponse(IEnumerable<Test> tests)
+    {
+        Tests = tests;
     }
 }
