@@ -1,13 +1,12 @@
 ﻿using System;
 using Xunit;
 
-namespace Peaky.XUnit
+namespace Peaky.XUnit;
+
+public class PeakyDataAttribute : ClassDataAttribute
 {
-    public class PeakyDataAttribute : ClassDataAttribute
+    public PeakyDataAttribute(Type testType) : base(testType)
     {
-        public PeakyDataAttribute(Type testType) : base(testType)
-        {
-        }
-        
     }
+        
 }
