@@ -1,9 +1,8 @@
 using System.Collections.Generic;
 
-namespace Peaky.XUnit.Tests.Model
+namespace Peaky.XUnit.Tests.Model;
+
+public class TestRun
 {
-    public class TestRun
-    {
-        public List<UnitTestResult> Results { get; set; }
-    }
+    public List<UnitTestResult> Results { get; set; }
 }
